*: fix some dumb printf format warnings
Some types like `time_t` vary across platforms and always need to be
cast when printed.

+++ lib/thread.c
@@ -96,8 +96,10 @@ static void vty_out_cpu_thread_history(struct vty *vty,
 				       struct cpu_thread_history *a)
 {
 	vty_out(vty, "%5zu %10zu.%03zu %9zu %8zu %9zu %8zu %9zu",
-		a->total_active, a->cpu.total / 1000, a->cpu.total % 1000,
-		a->total_calls, a->cpu.total / a->total_calls, a->cpu.max,
+		(size_t)a->total_active,
+		a->cpu.total / 1000, a->cpu.total % 1000,
+		(size_t)a->total_calls,
+		a->cpu.total / a->total_calls, a->cpu.max,
 		a->real.total / a->total_calls, a->real.max);
 	vty_out(vty, " %c%c%c%c%c %s\n",
 		a->types & (1 << THREAD_READ) ? 'R' : ' ',
